+++ src/main/scala/jobs/sessions/PurchasesPerSession.scala
@@ -69,7 +69,7 @@ class PurchasesPerSession(sc: SparkContext) extends Actor with ActorLogging  wit
       classOf[BSONObject]
     ).filter((t: Tuple2[Object, BSONObject]) => {
       def parseFloat(d: String): Option[Long] = {
-        try { Some(d.toLong) } catch { case _: Throwable => None }
+        try { Some(d.toDouble.toLong) } catch { case _: Throwable => None }
       }
       parseFloat(t._2.get("lowerDate").toString) match {
         case Some(dbDate) => {
